+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added Red block outline with wrong tool
   * If the tool in hand will make the block drop no items then the block outline will be green tinted
+* Berry bushes patches now have fewer bushes and no longer generate all fully grown
 * Fixed mobs arrows being affected by arrows damage multiplier
 
 ## 3.13.2