+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 3.13.3
 * Added Red block outline with wrong tool
   * If the tool in hand will make the block drop no items then the block outline will be green tinted
 * Added water push strength and push even if there are no blocks nearby